# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lorekeeper Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/campaigns": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "List Campaigns",
        "operationId": "list_campaigns",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampaignResponse"
                  },
                  "type": "array",
                  "title": "Response List Campaigns"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      },
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create Campaign",
        "operationId": "create_campaign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{slug}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaign",
        "operationId": "get_campaign",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Patch Campaign",
        "operationId": "patch_campaign",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCampaignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Delete Campaign",
        "operationId": "delete_campaign",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns/{slug}/chronicle/entries": {
      "get": {
        "tags": [
          "Chronicle"
        ],
        "summary": "List Chronicle Entries",
        "operationId": "list_chronicle_entries",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChronicleEntryResponse"
                  },
                  "title": "Response List Chronicle Entries"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Chronicle"
        ],
        "summary": "Create Chronicle Entry",
        "operationId": "create_chronicle_entry",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChronicleEntryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChronicleEntryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/api/v1/campaigns/{slug}/chronicle/entries/{entry_slug}": {
      "get": {
        "tags": [
          "Chronicle"
        ],
        "summary": "Get Chronicle Entry",
        "operationId": "get_chronicle_entry",
        "parameters": [
          {
            "name": "entry_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Slug"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChronicleEntryDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Chronicle"
        ],
        "summary": "Patch Chronicle Entry",
        "operationId": "patch_chronicle_entry",
        "parameters": [
          {
            "name": "entry_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Slug"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchChronicleEntryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChronicleEntryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      },
      "delete": {
        "tags": [
          "Chronicle"
        ],
        "summary": "Delete Chronicle Entry",
        "operationId": "delete_chronicle_entry",
        "parameters": [
          {
            "name": "entry_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Slug"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthorResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name"
        ],
        "title": "AuthorResponse",
        "example": {
          "display_name": "The Dungeon Master",
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        }
      },
...
      "CampaignResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          },
          "invite_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invite Code"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "slug",
          "role",
          "invite_code",
          "created_at",
          "updated_at"
        ],
        "title": "CampaignResponse",
        "example": {
          "created_at": "2024-01-15T10:00:00Z",
          "description": "A gothic horror campaign set in the mist-shrouded land of Barovia.",
          "id": "7c9e6679-7425-40de-944b-e07fc1f90ae7",
          "invite_code": "dX9kLmN2pQrS4tUvWxYz",
          "name": "Curse of Strahd",
          "role": "gm",
          "slug": "curse-of-strahd-a1b2c3d4",
          "updated_at": "2024-01-15T10:00:00Z"
        }
      },
...
      "ChronicleEntryDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "restricted": {
            "type": "boolean",
            "title": "Restricted"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "occurred_at",
          "body",
          "restricted",
          "tags",
          "created_at",
          "updated_at",
          "author"
        ],
        "title": "ChronicleEntryDetailResponse",
        "example": {
          "author": {
            "display_name": "The Dungeon Master",
            "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "body": "The party stormed the keep at dusk.",
          "created_at": "2024-01-16T02:30:00Z",
          "id": "a4c2b1d0-e3f5-4789-abcd-ef0123456789",
          "occurred_at": "2024-01-15T19:00:00Z",
          "restricted": false,
          "slug": "the-fall-of-blackspire",
          "tags": [
            "battle",
            "turning-point"
          ],
          "title": "The Fall of Blackspire",
          "updated_at": "2024-01-16T02:30:00Z"
        }
      },
      "ChronicleEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "restricted": {
            "type": "boolean",
            "title": "Restricted"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "occurred_at",
          "body",
          "restricted",
          "tags",
          "created_at",
          "updated_at"
        ],
        "title": "ChronicleEntryResponse",
        "example": {
          "body": "The party stormed the keep at dusk.",
          "created_at": "2024-01-16T02:30:00Z",
          "id": "a4c2b1d0-e3f5-4789-abcd-ef0123456789",
          "occurred_at": "2024-01-15T19:00:00Z",
          "restricted": false,
          "slug": "the-fall-of-blackspire",
          "tags": [
            "battle",
            "turning-point"
          ],
          "title": "The Fall of Blackspire",
          "updated_at": "2024-01-16T02:30:00Z"
        }
      },
      "CreateCampaignRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "slug_label": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*\\z",
            "title": "Slug Label"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug_label"
        ],
        "title": "CreateCampaignRequest",
        "example": {
          "description": "A gothic horror campaign set in the mist-shrouded land of Barovia.",
          "name": "Curse of Strahd",
          "slug_label": "curse-of-strahd"
        }
      },
...
      "CreateChronicleEntryRequest": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*\\z",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "restricted": {
            "type": "boolean",
            "title": "Restricted",
            "default": false
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title",
          "occurred_at"
        ],
        "title": "CreateChronicleEntryRequest",
        "example": {
          "body": "The party stormed the keep at dusk.",
          "occurred_at": "2024-01-15T19:00:00Z",
          "slug": "the-fall-of-blackspire",
          "title": "The Fall of Blackspire"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberRole": {
        "type": "string",
        "enum": [
          "gm",
          "player"
        ],
        "title": "MemberRole"
      },
      "PatchCampaignRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "slug_label": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*\\z",
            "title": "Slug Label"
          }
        },
        "type": "object",
        "title": "PatchCampaignRequest",
        "example": {
          "description": "Updated campaign description.",
          "name": "Curse of Strahd: Reloaded",
          "slug_label": "curse-of-strahd-reloaded"
        }
      },
...
      "PatchChronicleEntryRequest": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "restricted": {
            "type": "boolean",
            "title": "Restricted"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "PatchChronicleEntryRequest",
        "example": {
          "body": "Updated write-up.",
          "tags": [
            "battle",
            "turning-point"
          ],
          "title": "The Fall of Blackspire, Revised"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}